+++ src/main/java/learn/ang/umsback/service/AuthenticationService.java
@@ -1,0 +1,7 @@
+package learn.ang.umsback.service;
+
+import learn.ang.umsback.model.User;
+
+public interface AuthenticationService {
+    User signInAndReturnJwt(User signInRequest);
+}